app/src/backend/requirements.txt
torch; sys_platform == "darwin"
torchvision; sys_platform == "darwin"
pytorch-lightning; sys_platform == "darwin"


pillow>=9.0.0
imagehash>=4.3.1
flask==2.0.1
flask-cors==3.0.10
numpy>=1.21.0
pydantic>=1.8.2
transformers>=4.30.0
open-clip-torch>=2.20.0
chromadb
sentence-transformers
